Revert "main.goとわけた"
This reverts commit e5399876e950adefbdba8ebef44bb81b634f889d.
